split the worktree utils in a separate file
